chore: correction typos index.md
The standard library has it's own reference documentation that has detailed information about each module in std.

=>

The standard library has its own reference documentation that has detailed information about each module in std.

"it's" should be corrected to "its". "It's" is a contraction for "it is", while "its" is the possessive form.

Similar to ERC standards for Ethereum and Solidity, Sway has it's own SRC standards...

=>

Similar to ERC standards for Ethereum and Solidity, Sway has its own SRC standards...

"it's" should be corrected to "its".