Fixes ureachable macro crashing compiler.
In case of a bad input given to the lexer it is possible to generate
an AST with invalid ImplItem tuples.

This fix replaces the unreachable macro with an CompileError::Internal.
This allows the user to see the lexer errors and fix them which will
also take care of the new errors.

Fixes #6339.