Fix `impl self/trait` node order dependency pass key collision issue.
This fixes a key collision issue that happened in the node dependency
pass while handling `impl self/trait` nodes with multiple blocks.

See the following example for a simple example of the problem:

```
struct A {}

impl A {   //dependency of this impl is overwritten by the next impl
    fn a() {}
    fn b() {}
}

impl A {   //this will overwrite the dependency for the previous impl
    fn ab() {}
}
```

Fixes the problem by introducing a `,` while joining the names for the
key.

Closes https://github.com/FuelLabs/sway/issues/6319.