Fix trait-related type inference propagation bug.
Trait methods end up passing an extra type parameter related to self.

What was happening is that we ended up creating a placeholder of a
placeholder in some cases, which threw the type inferencing system off.

The fix is to avoid creating a new type for a given type parameter if
its already a placeholder.

Fixes https://github.com/FuelLabs/sway/issues/6372.