Fixes Numeric type propagation
To fix the Numeric type propagation we now do the type checking of code
blocks twice, first to collect all the unification and second to unify
the types of a previous namespace with the variable declarations types
of the current namespace.

This changes incurs a significative performance drop in the compilation time.
The test in release goes from 3 secs to 5 secs.

This can be improved significatively when we start using LexicalScopes
instead of cloning the whole namespace on each scope.

Fixes #6371