Fixes ABI alias metadata types.
Before this push using an alias in a contract would produce an ABI with
two distinct metadata types for tuples, one for `(_,_)` and another one
for `(u64,u64)`.

With this change alias are bypassed and we only produce metadata for the
string type of the inner alias type. If alias is `(u64, u64)` we will
produce a metadatatype for `(_,_)`.

Fixes #6488