Fix `function_cache` garbage collection bug  (#6555)
## Description
Fixes a bug that would crash the language server when a function node
was part of the AST.

#5967 added a function cache to the `QueryEngine`. Garbage collection
needed to be called on this otherwise it returns references to types
that have been cleared from the other engines during garbage collection.

In the future, any other nodes that we decide to cache need to have GC
applied to them or we will end up crashing the language server on the
first key-pressed event.

I've also removed the gc_frequency setting and now run GC on each
keystroke. Without doing this, the spans stored in the TokenMap and what
was returned from the compiler were falling out of sync. This ensures
that everything is always up to date and the correct spans are used.
closes #5260

Finally, I've added a `launch.json` that allows for attaching `lldb` to
a live running `forc-lsp` process. This was a pretty cruical step for
tracking down this bug so would be nice to have it easily accessible for
future debugging sessions.

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [ ] If my change requires substantial documentation changes, I have
[requested support from the DevRel
team](https://github.com/FuelLabs/devrel-requests/issues/new/choose)
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.

---------

Co-authored-by: Sophie Dankel <47993817+sdankel@users.noreply.github.com>