fix: forc-client build script should update proxy abi also for deployment tests (#6561)
## Description
closes #6558.

This PR fixes an issue with the forc-client build script that causes the
deployment tests to be updated manually for each proxy abi change. This
is a follow up on #6535 and was overlooked in that one. Not critical as
it does not break anything on the user side but rather an code-quality
issue.