Fix module ancestor check for imports (#6671)
## Description

When importing items we allow imports from ancestor modules, even if the
ancestor module is declared as private.

The code that performs this check was moved from `module.rs` to
`root.rs` as part of #5697. During this move the current module path
(i.e., the path of the destination of the import) was mistakenly
replaced with the module path of the root module of the destination
module. Since root modules do not have ancestors this meant that no
source module was ever considered an ancestor of the destination module.

This PR fixes this problem, and adds an abscure test case that shows the
issue.


## Checklist

- [ ] I have linked to any relevant issues.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [ ] If my change requires substantial documentation changes, I have
[requested support from the DevRel
team](https://github.com/FuelLabs/devrel-requests/issues/new/choose)
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [ ] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [ ] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [ ] I have requested a review from the relevant team or maintainers.

---------

Co-authored-by: Joshua Batty <joshpbatty@gmail.com>