chore: remove unmaintained `ansi_term` in favor of `ansiterm` (#6696)
## Description

related to #2601.

Removes ansi_term dependency from sway crates. To remove the dependency
completely from dep tree we need a release of sway repo to get a new
version of forc_util and use it in forc_wallet as forc_wallet is
depending on `forc-util v0.47.0` which uses `ansi_term`. Once that is
done we can close #2601