Fix handling of initialized associated constants in ABIs and traits.
From the reporting issue:

Current implementation of the initialized associated constants in ABIs
and traits is contradictory. Namely, even if we initialize them, we
anyway always have to implement them in the contract or type `impl`s.
This eliminates the whole purpose of being able to initialize them on
the ABI/trait level.

E.g., for ABIs (and the same is for traits):

```
abi ABI {
const CONST: u32 = 111; // <<<--- Initialized here. But this value
cannot be used, because...
}

impl ABI for Contract {
// ... associated constants MUST always be implemented in the
contract/type and initialized anew.
const CONST: u32 = 222; // <<<--- This must be here and this value
wins.
}
```

What we want to have instead, is the same behavior as in Rust. If an
associated constant is already initialized in the ABI or trait
declaration, the `impl`s do not need to provide the constant. If they
do, the value from the `impl` wins. The existing checks for having the
same constant type in the ABI/trait declaration and `impl`s, of course,
remain.

Fixes https://github.com/FuelLabs/sway/issues/6343.