correct text index.md
The list of examples indicates twice a link to Wallet_smart_contract.md, instead, one of the links should be conducted on Liquidity_pool.md. I replaced the link to Liquidity_pool.md instead of Wallet_Smart_contract.md
FizzBuzz  had not a correct style as he was in quotation marks. The style is changed