Function selector collision check (#6869)
## Description

Fixes #6334 .

When compiling a contract call using v0 encoding the compiled code uses
a 4 byte function selector to determine which method is to be called.
This means that all methods in a contract must have distinct function
selectors to ensure that the wrong method isn't called by accident.

We have so far not had a check for distinctness of function selectors.
This PR introduces such a check.

Ideally the check should be performed on the ABI declaration rather than
on the contract, but since the method `to_function_selector_value` is
not implemented for trait methods (which are used to represent ABI
methods) I have opted to perform the check on contract declarations
instead. It would probably be possible to implement
`to_functiion_selector_value` on trait methods, but since we are in the
process of switching to v1 encoding it didn't seem worth it to spend
time on that.

This change may cause existing contracts to no longer compile, but in
those cases the contract's ABI declaration is unsafe, so I do not
consider this to be a breaking change.


## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] If my change requires substantial documentation changes, I have
[requested support from the DevRel
team](https://github.com/FuelLabs/devrel-requests/issues/new/choose)
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.