fix: remove invalid unpacking logic for IPFS pinned pkg sources (#6902)
## Description
closes #6721.

There were couple of issues:

1. Archive unpacking was getting the bytes from `text` field of response
to the IPFS gateway which was adding additional stuff that messes with
unpacking routine.
2. Folder name was invalid so forc was unable to find the packages in
the `ipfs` cache.

Both issues are addressed and this unblocks forc to fetch ipfs fetched
packages once again. Next up on our package registry push is to enable
alternative sources and set the precedence for them.

Also added some tests around extracting logic, which inserts the
contents under a directory named after `CID` as this is how `forc` is
looking for these sources in the first place (related to point 2 above)