fix extract_type_parameters regarding primitive types (#6965)
## Description

This PR is part of https://github.com/FuelLabs/sway/issues/5110. It
fixes a problem that improves the usability of slices.

When a method is returned from the "method resolution algorithm", if any
generic is involved, it is returned as its generic form. So for example,
if a method `len` is called on `Vec<u8>`, we return something like `fn
len(self: Vec<T>)`.

Later we monomorphize `T` into `u8`. To do that we run a "pattern match"
to realize that we need to replace `T` with `u8`.

In this case, the bug was that we were not considering that some
primitive types have "implicit type parameters".

This was found whilst creating more tests for
https://github.com/FuelLabs/sway/pull/6933, because I was not able to
call the `len` method defined for slices as:

```sway
impl<T> &[T] {
    pub fn len(self) -> u64 {
        ...
    }
}
``` 

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [ ] If my change requires substantial documentation changes, I have
[requested support from the DevRel
team](https://github.com/FuelLabs/devrel-requests/issues/new/choose)
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.