ci: add a step to reject x.y.z versioning for dependencies (#7082)
## Description

closes #7035.

Adds a CI step to ensure we do not introduce x.y.z versions in the
`Cargo.toml`. For fuel-core-client we enforce the version to be in form
x.y.z. For some pre-defined set we allow either x.y.z and x.y
(`etk-asm`, `etk-ops`, `dap`) and for the rest we enforce that the
versions are in the form of x.y